+++ db-dumps/dump.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-pg_dump -h localhost -U postgres -W l22 > `date +l22-%F.sql`
+pg_dump -h localhost -U postgres -W qa > `date +qa-%F.sql`