+++ src/qa/boundary/users.clj
@@ -1,18 +1,7 @@
 (ns qa.boundary.users
- (:require
-  ;;[environ.core :refer [env]]
-  ;;[next.jdbc :refer [get-connection]]
-  ;;[next.jdbc.result-set :as rs]
-  [next.jdbc.sql :refer [query]]
-  [qa.boundary.utils :refer [ds-opt]]))
-
-;; (def db {:dbtype   "postgresql"
-;;          :host     (env :l22-host)
-;;          :dbname   (env :l22-db)
-;;          :user     (env :l22-user)
-;;          :password (env :l22-password)})
-
-;;(def ds (get-connection db))
+  (:require
+   [next.jdbc.sql :refer [query]]
+   [qa.boundary.utils :refer [ds-opt]]))
 
 ;; typing から持ってきた関数名をそのまま。
 (defn find-user-by-login [db login]